[NEAR-119] Refactor: concert 분리에 따른 import 경로 변경
