+++ app/api/router.py
@@ -3,10 +3,10 @@ from fastapi import APIRouter
 from app.domains.artists.router import router as artists_router
 from app.domains.auth.router import router as auth_router
 from app.domains.chat.router_ws import router as chat_router
+from app.domains.concerts.router import router as concert_admin_router
 from app.domains.notifications.router import router as notifications_router
-from app.domains.streams.admin.concert_router import router as concert_admin_router
 from app.domains.streams.admin.ivs_webhook_router import router as streams_status_router
-from app.domains.streams.admin.stream_router import router as streams_admin_router
+from app.domains.streams.admin.router import router as streams_admin_router
 from app.domains.streams.client.router import router as streams_router
 from app.domains.users.router import router as users_router
